+++ src/main/java/com/treasure_data/td_import/BulkImportCommand.java
@@ -209,8 +209,8 @@ public final class BulkImportCommand extends BulkImport {
         // 'auto-perform' and 'auto-commit'
         results.add(UploadProcessor.processAfterUploading(biClient, uploadConf, sessionName));
 
-        if (hasNoUploadError(results) && uploadConf.autoDelete()) {
-            // 'auto-delete-session'
+        // 'auto-delete'
+        if (hasNoUploadError(results) && uploadConf.autoDelete() && uploadConf.autoCommit()) {
             results.add(UploadProcessor.deleteSession(biClient, uploadConf, sessionName));
         }
